requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
google-cloud-vision
google-cloud-pubsub
google-generativeai
google-adk
python-dotenv
requests
gunicorn
pydantic
asyncio-extras
